refactor: utils files
